requirements.txt
# Flask + MySQL - DW Manager
Flask>=3.0.0
PyMySQL>=1.1.0
bcrypt>=4.0.0
python-dotenv>=1.0.0
